backend/requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
promise==2.3
psycopg2==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
requests==2.28.1
simplejson==3.17.6
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.10
whitenoise==6.2.0
